Moved experimental API endpoints to separate module
