+++ sciety_labs/app/routers/api/app.py
@@ -8,6 +8,7 @@ from sciety_labs.app.routers.api.article_recommendation import (
     create_api_article_recommendation_router
 )
 from sciety_labs.app.routers.api.debug import create_api_debug_router
+from sciety_labs.app.routers.api.experimental import create_api_experimental_router
 
 
 def create_api_app(
@@ -20,6 +21,9 @@ def create_api_app(
         app_update_manager=app_update_manager
     ))
     app.include_router(create_api_debug_router())
+    app.include_router(create_api_experimental_router(
+        app_providers_and_models=app_providers_and_models
+    ))
     app.include_router(create_api_article_recommendation_router(
         app_providers_and_models=app_providers_and_models
     ))
