+++ sciety_labs/app/routers/api/article_recommendation.py
@@ -396,91 +396,4 @@ def create_api_article_recommendation_router(
             fields=fields_set
         )
 
-    @router.get(
-        '/experimental/opensearch/metadata/by/doi'
-    )
-    async def experimental_opensearch_metadata_by_doi(
-        request: fastapi.Request,
-        article_doi: str = fastapi.Query(
-            description=textwrap.dedent(
-                '''
-                The DOI
-                '''
-            )
-        )
-    ):
-        doc = await (
-            app_providers_and_models
-            .async_opensearch_client
-            .get_source(
-                index=app_providers_and_models.opensearch_config.index_name,
-                id=article_doi,
-                headers=get_cache_control_headers_for_request(request)
-            )
-        )
-        return doc
-
-    @router.get(
-        '/experimental/sync/crossref/metadata/by/doi'
-    )
-    def experimental_sync_crossref_metadata_by_doi(
-        request: fastapi.Request,
-        article_doi: str
-    ):
-        return (
-            app_providers_and_models
-            .crossref_metadata_provider
-            .get_crossref_metadata_dict_by_doi(
-                article_doi,
-                headers=get_cache_control_headers_for_request(request)
-            )
-        )
-
-    @router.get(
-        '/experimental/async/crossref/metadata/by/doi'
-    )
-    async def experimental_async_crossref_metadata_by_doi(
-        request: fastapi.Request,
-        article_doi: str
-    ):
-        return await (
-            app_providers_and_models
-            .async_crossref_metadata_provider
-            .get_crossref_metadata_dict_by_doi(
-                article_doi,
-                headers=get_cache_control_headers_for_request(request)
-            )
-        )
-
-    @router.get(
-        '/experimental/embedding-vector-for-title-abstract'
-    )
-    async def experimental_embedding_vector_for_title_abstract(
-        request: fastapi.Request,
-        title: str = fastapi.Query(
-            description=textwrap.dedent(
-                '''
-                The title
-                '''
-            )
-        ),
-        abstract: str = fastapi.Query(
-            description=textwrap.dedent(
-                '''
-                The abstract
-                '''
-            )
-        )
-    ):
-        embedding_vector = await (
-            app_providers_and_models
-            .async_title_abstract_embedding_vector_provider
-            .get_embedding_vector(
-                title=title,
-                abstract=abstract,
-                headers=get_cache_control_headers_for_request(request)
-            )
-        )
-        return embedding_vector
-
     return router
